several build changes
- separated MEKA stuff into jaicore-ml-meka
- moved meta-learning stuff to mlplan-metalearning project
- removed nd4j and optimizer libraries from the ml-plan dependencies,
which reduces the size of far jars from 700MB to under 100MB

+++ softwareconfiguration/mlplan/core/src/main/java/ai/libs/mlplan/core/MLPlan.java
@@ -289,7 +289,7 @@ public class MLPlan<L extends ISupervisedLearner<ILabeledInstance, ILabeledDatas
 		this.logger.info("Switched ML-Plan logger to {}", name);
 	}
 
-	public void setPortionOfDataForPhase2(final float portion) {
+	public void setPortionOfDataForPhase2(final double portion) {
 		this.getConfig().setProperty(MLPlanClassifierConfig.SELECTION_PORTION, String.valueOf(portion));
 	}
